move registration patch
